+++ src/api/general-services/pipeline-manage/constructors/construct-pipeline-step.js
@@ -3,7 +3,7 @@ const createNewStep = require('./create-new-step');
 const createNewJobIfNotExist = require('./create-new-job-if-not-exist');
 const {
   requestPod, waitForPod,
-} = require('./assign-pod-to-pipeline');
+} = require('./assignPodToPipeline');
 
 const constructPipelineStep = (context, step) => {
   const { XStepType: stepType, XConstructorArgs: args } = step;
