+++ tests/api/general-services/state-machine-constructors.test.js
@@ -1,6 +1,6 @@
 const constants = require('../../../src/api/general-services/pipeline-manage/constants');
 const fake = require('../../test-utils/constants');
-const { buildPodRequest } = require('../../../src/api/general-services/pipeline-manage/constructors/assign-pod-to-pipeline');
+const { buildPodRequest } = require('../../../src/api/general-services/pipeline-manage/constructors/assignPodToPipeline');
 const validateRequest = require('../../../src/utils/schema-validator');
 
 describe('Test for pipeline constructor services', () => {
