+++ tests/api/route-services/pipeline-response.test.js
@@ -4,7 +4,7 @@ const io = require('socket.io-client');
 const validateRequest = require('../../../src/utils/schema-validator');
 const PipelineService = require('../../../src/api/route-services/pipeline-response');
 const pipelineAssign = require('../../../src/utils/hooks/pipeline-assign');
-const { buildPodRequest } = require('../../../src/api/general-services/pipeline-manage/constructors/assign-pod-to-pipeline');
+const { buildPodRequest } = require('../../../src/api/general-services/pipeline-manage/constructors/assignPodToPipeline');
 const constants = require('../../../src/api/general-services/pipeline-manage/constants');
 const { SUCCEEDED } = require('../../../src/api/general-services/pipeline-manage/constants');
 const fake = require('../../test-utils/constants');
