+++ tests/api/utils/hooks/pipeline-assign.test.js
@@ -2,7 +2,7 @@
 const k8s = require('@kubernetes/client-node');
 const constants = require('../../../../src/api/general-services/pipeline-manage/constants');
 const fake = require('../../../test-utils/constants');
-const { buildPodRequest } = require('../../../../src/api/general-services/pipeline-manage/constructors/assign-pod-to-pipeline');
+const { buildPodRequest } = require('../../../../src/api/general-services/pipeline-manage/constructors/assignPodToPipeline');
 
 jest.mock('@kubernetes/client-node');
